requirements.txt
pip
numpy
pandas
streamlit
requests
st-annotated-text
transformers
keras
tensorflow
io

